requirements.txt
python-telegram-bot==20.7
psutil==5.9.6
requests==2.31.0
python-dotenv==1.2.1